\name{CRFutilRcppComponents-package}
\alias{CRFutilRcppComponents-package}
\alias{CRFutilRcppComponents}
\docType{package}
\title{
\packageTitle{CRFutilRcppComponents}
}
\description{
\packageDescription{CRFutilRcppComponents}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CRFutilRcppComponents}
\packageIndices{CRFutilRcppComponents}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CRFutilRcppComponents}

Maintainer: \packageMaintainer{CRFutilRcppComponents}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
